requirements.txt
tensorflow-gpu>=2.5.0
setuptools==41.0.1
ftfy==5.6
tqdm==4.32.1
Click==7.0






